requirements.txt
annotated-types==0.5.0
anyio==3.7.1
cffi==1.15.1
click==8.1.7
cryptography==41.0.3
exceptiongroup==1.1.3
fastapi==0.103.1
h11==0.14.0
idna==3.4
peewee==3.16.3
pycparser==2.21
pydantic==2.3.0
pydantic-core==2.6.3
PyMySQL==1.1.0
PyYAML==6.0.1
sniffio==1.3.0
starlette==0.27.0
typing-extensions==4.7.1
uvicorn==0.23.2
